Short (50 char or less) summary of changes
More detail explanatory text, if necessary.  Wrap it to about 72
characters or so.  In some contexts, the first line is treated as the
subject of an email and the rest of the text is the body.  The blank
line is separating the summary from the body is critical (unless you
omit the body entirely); tools like rebase can get confused if you run
the two together.

Further paragraphs come after blank lines.

  * Bullet points are okay, too.
  * Typically a hyphen or asterisk is used
    for the bullet, preceded by a single space, with blank likes in
    between but conventions vary here.