Updated the configuration to work with MacOS
Well only updated it to work with MacOS, partially. The configuration
needs some work to be right, as there appears to be an issue with the
use of spaces in path names (at leas for the Hoard configuration file).

The next step is to add in the configuration for NeoVim for consistency
across all environments.